+++ tests/unit/test__helpers.py
@@ -672,6 +672,7 @@ class Test_metadata_with_prefix(unittest.TestCase):
         self.assertEqual(metadata, [("google-cloud-resource-prefix", prefix)])
 
 
+
 class Test_retry(unittest.TestCase):
     class test_class:
         def test_fxn(self):
@@ -747,3 +748,16 @@ class Test_retry(unittest.TestCase):
         )
 
         self.assertEqual(test_api.test_fxn.call_count, 3)
+
+class Test_metadata_with_leader_aware_routing(unittest.TestCase):
+    def _call_fut(self, *args, **kw):
+        from google.cloud.spanner_v1._helpers import _metadata_with_leader_aware_routing
+
+        return _metadata_with_leader_aware_routing(*args, **kw)
+
+    def test(self):
+        value = True
+        metadata = self._call_fut(True)
+        self.assertEqual(
+            metadata, ("x-goog-spanner-route-to-leader", str(value).lower())
+        )