+++ tests/unit/test_snapshot.py
@@ -1108,7 +1108,10 @@ class Test_SnapshotBase(OpenTelemetryBase):
         )
         api.partition_read.assert_called_once_with(
             request=expected_request,
-            metadata=[("google-cloud-resource-prefix", database.name)],
+            metadata=[
+                ("google-cloud-resource-prefix", database.name),
+                ("x-goog-spanner-route-to-leader", "true"),
+            ],
             retry=retry,
             timeout=timeout,
         )
@@ -1279,7 +1282,10 @@ class Test_SnapshotBase(OpenTelemetryBase):
         )
         api.partition_query.assert_called_once_with(
             request=expected_request,
-            metadata=[("google-cloud-resource-prefix", database.name)],
+            metadata=[
+                ("google-cloud-resource-prefix", database.name),
+                ("x-goog-spanner-route-to-leader", "true"),
+            ],
             retry=retry,
             timeout=timeout,
         )
@@ -1744,6 +1750,7 @@ class _Database(object):
     def __init__(self):
         self.name = "testing"
         self._instance = _Instance()
+        self._route_to_leader_enabled = True
 
 
 class _Session(object):